Last fixes for lottie support
