+++ packages/teleport-uidl-resolver/__tests__/embed-lottie-node/index.ts
@@ -1,8 +1,4 @@
-import {
-  wrapHtmlLottieNode,
-  createEmbedDivWrapperNode,
-  createLottieDivWrapperNode,
-} from '../../src/resolvers/embed-lottie-node/utils'
+import { wrapHtmlNode, createEmbedDivWrapperNode } from '../../src/resolvers/embed-node/utils'
 import { elementNode, staticNode } from '@teleporthq/teleport-uidl-builders'
 import { UIDLElementNode, UIDLStyleDefinitions } from '@teleporthq/teleport-types'
 import { create } from 'domain'
@@ -24,7 +20,7 @@ describe('wrap html-node element', () => {
       }
     )
 
-    const result = wrapHtmlLottieNode(node, {})
+    const result = wrapHtmlNode(node, {})
     expect(result.content.elementType).toBe('div')
     expect(result.content.children?.length).toBe(1)
     const childNode = result.content.children?.[0] as UIDLElementNode
@@ -59,86 +55,5 @@ describe('create div', () => {
     expect(result.content.attrs?.html).toBeUndefined()
     expect(result.content.style).toBeDefined()
     expect(Object.keys(result.content.style as UIDLStyleDefinitions).length).toBe(1)
-  })
-})
-
-describe('wrap lottie element', () => {
-  it('wraps a simple element', () => {
-    const node = elementNode(
-      'lottie-node',
-      {
-        src: {
-          type: 'static',
-          content: 'https://assets9.lottiefiles.com/datafiles/gUENLc1262ccKIO/data.json',
-        },
-        autoplay: {
-          type: 'static',
-          content: 'true',
-        },
-        id: {
-          type: 'static',
-          content: 'id_123',
-        },
-      },
-      [],
-      undefined,
-      {
-        width: staticNode('100px'),
-        height: staticNode('100px'),
-      }
-    )
-
-    const result = wrapHtmlLottieNode(node, {})
-    expect(result.content.elementType).toBe('div')
-    expect(result.content.children?.length).toBe(1)
-    expect(result.content.children?.length).toBe(1)
-    expect(result.content.style?.width.content).toBe('100px')
-    expect(result.content.style?.height.content).toBe('100px')
-    expect(result.content.attrs?.id.content).toBe('id_123')
-
-    const childNode = result.content.children?.[0] as UIDLElementNode
-    expect(childNode.content.elementType).toBe('lottie-node')
-    expect(childNode.content.style).toStrictEqual({})
-    expect(childNode.content.attrs?.src.content).toBe(
-      `https://assets9.lottiefiles.com/datafiles/gUENLc1262ccKIO/data.json`
-    )
-    expect(childNode.content.attrs?.autoplay.content).toBe('true')
-  })
-})
-
-describe('create div', () => {
-  it('creates wrapping div for lottie node', () => {
-    const node = elementNode(
-      'lottie-node',
-      {
-        src: {
-          type: 'static',
-          content: 'https://assets9.lottiefiles.com/datafiles/gUENLc1262ccKIO/data.json',
-        },
-        autoplay: {
-          type: 'static',
-          content: 'true',
-        },
-        id: {
-          type: 'static',
-          content: 'id_123',
-        },
-      },
-      [],
-      undefined,
-      {
-        width: staticNode('100px'),
-        height: staticNode('100px'),
-      }
-    )
-
-    const result = createLottieDivWrapperNode(node)
-
-    expect(result.content.elementType).toBe('div')
-    expect(result.content.children?.length).toBe(0)
-    expect(result.content.attrs?.src).toBeUndefined()
-    expect(result.content.style?.width.content).toBe('100px')
-    expect(result.content.style?.height.content).toBe('100px')
-    expect(result.content.attrs?.id.content).toBe('id_123')
   })
 })
