+++ packages/teleport-uidl-resolver/src/resolvers/embed-node/index.ts
@@ -1,6 +1,6 @@
 import { ComponentUIDL, GeneratorOptions } from '@teleporthq/teleport-types'
-import { wrapHtmlLottieNode } from './utils'
+import { wrapHtmlNode } from './utils'
 
 export const resolveHtmlNode = (uidl: ComponentUIDL, options: GeneratorOptions) => {
-  uidl.node = wrapHtmlLottieNode(uidl.node, options)
+  uidl.node = wrapHtmlNode(uidl.node, options)
 }
